+++ pycm/__main__.py
@@ -3,8 +3,8 @@
 
 import doctest
 import sys
-from .pycm_obj import *
-from .pycm_output import *
+from .cm import *
+from .output import *
 from art import tprint
 
 
@@ -13,7 +13,7 @@ if __name__ == "__main__":
     if len(args) > 1:
         if args[1].upper() == "TEST":
             error_flag = doctest.testfile(
-                "pycm_test.py",
+                "basic_test.py",
                 optionflags=doctest.NORMALIZE_WHITESPACE | doctest.ELLIPSIS
                 | doctest.IGNORE_EXCEPTION_DETAIL,
                 verbose=False)[0]