+++ pycm/class_funcs.py
@@ -2,9 +2,9 @@
 """Class statistics functions."""
 from __future__ import division
 import math
-from .pycm_util import normal_quantile
-from .pycm_interpret import *
-from .pycm_param import CLASS_PARAMS
+from .utils import normal_quantile
+from .interpret import *
+from .params import CLASS_PARAMS
 
 
 def sensitivity_index_calc(TPR, FPR):