+++ pycm/cm.py
@@ -1,17 +1,17 @@
 # -*- coding: utf-8 -*-
 """ConfusionMatrix module."""
 from __future__ import division
-from .pycm_error import pycmVectorError, pycmMatrixError, pycmCIError, pycmAverageError, pycmPlotError
-from .pycm_handler import __class_stat_init__, __overall_stat_init__
-from .pycm_handler import __obj_assign_handler__, __obj_file_handler__, __obj_matrix_handler__, __obj_vector_handler__, __obj_array_handler__
-from .pycm_handler import __imbalancement_handler__
-from .pycm_class_func import F_calc, IBA_calc, TI_calc, NB_calc, sensitivity_index_calc
-from .pycm_overall_func import weighted_kappa_calc, weighted_alpha_calc, alpha2_calc, brier_score_calc, log_loss_calc
-from .pycm_distance import DistanceType, DISTANCE_MAPPER
-from .pycm_output import *
-from .pycm_util import *
-from .pycm_param import *
-from .pycm_ci import __CI_overall_handler__, __CI_class_handler__
+from .errors import pycmVectorError, pycmMatrixError, pycmCIError, pycmAverageError, pycmPlotError
+from .handlers import __class_stat_init__, __overall_stat_init__
+from .handlers import __obj_assign_handler__, __obj_file_handler__, __obj_matrix_handler__, __obj_vector_handler__, __obj_array_handler__
+from .handlers import __imbalancement_handler__
+from .class_funcs import F_calc, IBA_calc, TI_calc, NB_calc, sensitivity_index_calc
+from .overall_funcs import weighted_kappa_calc, weighted_alpha_calc, alpha2_calc, brier_score_calc, log_loss_calc
+from .distance import DistanceType, DISTANCE_MAPPER
+from .output import *
+from .utils import *
+from .params import *
+from .ci import __CI_overall_handler__, __CI_class_handler__
 import os
 import json
 import numpy