+++ pycm/compare.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 """Compare module."""
 from __future__ import division
-from .pycm_error import pycmCompareError
-from .pycm_output import *
-from .pycm_util import *
-from .pycm_param import *
-from .pycm_obj import ConfusionMatrix
+from .errors import pycmCompareError
+from .output import *
+from .utils import *
+from .params import *
+from .cm import ConfusionMatrix
 import os
 import numpy
 from warnings import warn