+++ pycm/handlers.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 """ConfusionMatrix handlers."""
 from __future__ import division
-from .pycm_class_func import class_statistics
-from .pycm_error import pycmVectorError, pycmMatrixError
-from .pycm_overall_func import overall_statistics
-from .pycm_util import *
-from .pycm_param import *
+from .class_funcs import class_statistics
+from .errors import pycmVectorError, pycmMatrixError
+from .overall_funcs import overall_statistics
+from .utils import *
+from .params import *
 import json
 import types
 import numpy