+++ pycm/multilabel_cm.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 """MultiLabelCM module."""
 from __future__ import division
-from .pycm_error import pycmVectorError, pycmMultiLabelError
-from .pycm_param import *
-from .pycm_obj import ConfusionMatrix
+from .errors import pycmVectorError, pycmMultiLabelError
+from .params import *
+from .cm import ConfusionMatrix
 import numpy