+++ pycm/overall_funcs.py
@@ -4,9 +4,9 @@ from __future__ import division
 import math
 import operator as op
 from functools import reduce
-from .pycm_interpret import *
-from .pycm_ci import kappa_SE_calc, CI_calc, SE_calc
-from .pycm_util import complement
+from .interpret import *
+from .ci import kappa_SE_calc, CI_calc, SE_calc
+from .utils import complement
 
 
 def log_loss_calc(